+++ src/models/BaseFieldProps.ts
@@ -3,5 +3,6 @@
 export interface BaseFieldProps<T> {
   label: string;
   value: T | null;
+  description?: string;
   required?: boolean;
 }